Erstattet TilkjentYtelseUtil med tre ulike servicer: TilkjentYtelseService, OppdaterAndelerMedEndretUtbetalingService og RegelverkFørFebruar2025AndelGeneratorGammel
