+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/beregning/TilkjentYtelseService.kt
@@ -1,0 +1,45 @@
+package no.nav.familie.ks.sak.kjerne.beregning
+
+import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkårsvurdering
+import no.nav.familie.ks.sak.kjerne.beregning.domene.TilkjentYtelse
+import no.nav.familie.ks.sak.kjerne.beregning.endretUtbetaling.OppdaterAndelerMedEndretUtbetalingService
+import no.nav.familie.ks.sak.kjerne.beregning.regelverkFørFebruar2025.gammel.RegelverkFørFebruar2025AndelGeneratorGammel
+import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonType
+import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonopplysningGrunnlag
+import org.springframework.stereotype.Service
+import java.time.LocalDate
+
+@Service
+class TilkjentYtelseService(
+    private val oppdaterAndelerMedEndretUtbetalingService: OppdaterAndelerMedEndretUtbetalingService,
+    private val regelverkFørFebruar2025AndelGeneratorGammel: RegelverkFørFebruar2025AndelGeneratorGammel,
+) {
+    fun beregnTilkjentYtelse(
+        vilkårsvurdering: Vilkårsvurdering,
+        personopplysningGrunnlag: PersonopplysningGrunnlag,
+        endretUtbetalingAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse> = emptyList(),
+    ): TilkjentYtelse {
+        val tilkjentYtelse =
+            TilkjentYtelse(
+                behandling = vilkårsvurdering.behandling,
+                opprettetDato = LocalDate.now(),
+                endretDato = LocalDate.now(),
+            )
+        val endretUtbetalingAndelerBarna = endretUtbetalingAndeler.filter { it.person?.type == PersonType.BARN }
+
+        val andelerTilkjentYtelseBarnaUtenEndringer =
+            regelverkFørFebruar2025AndelGeneratorGammel.beregnAndelerTilkjentYtelseForBarna(
+                personopplysningGrunnlag = personopplysningGrunnlag,
+                vilkårsvurdering = vilkårsvurdering,
+                tilkjentYtelse = tilkjentYtelse,
+            )
+
+        val andelerTilkjentYtelseBarnaMedAlleEndringer =
+            oppdaterAndelerMedEndretUtbetalingService.oppdaterTilkjentYtelseMedEndretUtbetalingAndeler(
+                andelTilkjentYtelserUtenEndringer = andelerTilkjentYtelseBarnaUtenEndringer,
+                endretUtbetalingAndeler = endretUtbetalingAndelerBarna,
+            )
+        tilkjentYtelse.andelerTilkjentYtelse.addAll(andelerTilkjentYtelseBarnaMedAlleEndringer.map { it.andel })
+        return tilkjentYtelse
+    }
+}
